Modified the http logging routine in order to catch arbitrary large chunks with arbitrary characters inside (also \r\n\r\n etc.)

+++ src/logger.h
@@ -40,8 +40,8 @@ struct http_client_struct {
 	char method[20];
 	char requestHeader[MAXLINE];
 	char requestBodyBinaryLocation[1000];
-        u_int64_t sent_content_length; // the conteht length of the whole chunk of data we expect to send
-	u_int64_t sent_content_done;
+        u_int32_t sent_content_length; // the conteht length of the whole chunk of data we expect to send
+	u_int32_t sent_content_done;
 	char* sent_content_done_ptr; // pointer to memory where we store the already sent data
 };
 
@@ -51,8 +51,8 @@ struct http_server_struct {
 	char responseLastModified[MAXLINE];
 	char responseContentType[1000];
 	char serverType[1000];
-	u_int64_t rcvd_content_length; // the content length of the whole chunk of data we expect to receive 
-	u_int64_t rcvd_content_done;
+	u_int32_t rcvd_content_length; // the content length of the whole chunk of data we expect to receive 
+	u_int32_t rcvd_content_done;
 	char* rcvd_content_done_ptr; // pointer to memory where we store the already received data
 };
